Switch select instructions over to using inline assembly
This choice was made in the name of efficiency and laziness. Efficiency
because many instructions have a direct equivalent in x86, e.g. ROR,
and their equivalent C++ implementation would be many times large than
a single instruction. Laziness for the simple reason that I don't like
calculating status flags.

The switched instructions are:
-ROR
-ROL
-ASR
-ASL
-SWAB